+++ lib/src/kuzzle/role.dart
