+++ lib/src/kuzzle/user.dart
